refactor(dashboard): consolidate scattered time formatting into lib/format-time
5 components had local time formatting functions that duplicated
lib/format-time.ts utilities:
- keeper-fleet-overview: formatRecency -> formatElapsed
- telemetry-unified: timeAgo -> formatTimeAgo (was already a wrapper)
- autoresearch: formatTimestamp/formatUpdatedAt -> formatTimestampKo
- connector-status: formatUptime/timeAgo -> formatElapsedCompact/formatTimeAgoEn
- server-config: formatUptime -> formatElapsedCompact

Added formatTimeAgoEn to lib/format-time for English relative time.
Net: -19 lines, 5 fewer local definitions.
