+++ dashboard/src/components/server-config.ts
@@ -6,6 +6,7 @@ import { TransportHealthPanel } from './transport-health'
 import { fetchDashboardConfig } from '../api/dashboard'
 import type { DashboardConfigResponse, ConfigEntry } from '../api/dashboard'
 import { createAsyncResource } from '../lib/async-state'
+import { formatElapsedCompact } from '../lib/format-time'
 
 const configResource = createAsyncResource<DashboardConfigResponse>()
 const searchQuery = signal('')
@@ -29,14 +30,8 @@ function toggleCategory(name: string) {
   expandedCategories.value = next
 }
 
-function formatUptime(seconds: number): string {
-  const h = Math.floor(seconds / 3600)
-  const m = Math.floor((seconds % 3600) / 60)
-  const s = Math.floor(seconds % 60)
-  if (h > 0) return `${h}h ${m}m ${s}s`
-  if (m > 0) return `${m}m ${s}s`
-  return `${s}s`
-}
+// Delegated to lib/format-time (SSOT)
+const formatUptime = formatElapsedCompact
 
 function matchesSearch(entry: ConfigEntry, query: string): boolean {
   if (!query) return true
