+++ dashboard/src/lib/format-time.ts
@@ -80,6 +80,18 @@ export function formatTimestampKo(ts: number): string {
   })
 }
 
+/** Format ISO timestamp as English relative time — "just now", "5m ago", "2h ago". */
+export function formatTimeAgoEn(iso: string): string {
+  if (!iso.trim()) return 'unknown'
+  if (iso === 'never') return 'never'
+  const diff = (Date.now() - new Date(iso).getTime()) / 1000
+  if (Number.isNaN(diff)) return 'unknown'
+  if (diff < 60) return 'just now'
+  if (diff < 3600) return `${Math.floor(diff / 60)}m ago`
+  if (diff < 86400) return `${Math.floor(diff / 3600)}h ago`
+  return `${Math.floor(diff / 86400)}d ago`
+}
+
 /** Format a numeric delta with sign prefix. */
 export function formatDelta(delta: number, decimals = 4): string {
   const sign = delta >= 0 ? '+' : ''
